Move generic dispatch functions to dispatch.rs
